+++ app/src/main/java/pl/lambada/songsync/data/remote/lyrics_providers/LyricsProviderService.kt
@@ -7,10 +7,10 @@ import pl.lambada.songsync.data.remote.lyrics_providers.others.NeteaseAPI
 import pl.lambada.songsync.data.remote.lyrics_providers.spotify.SpotifyAPI
 import pl.lambada.songsync.data.remote.lyrics_providers.spotify.SpotifyLyricsAPI
 import pl.lambada.songsync.domain.model.SongInfo
-import pl.lambada.songsync.ui.screens.Providers
-import pl.lambada.songsync.ui.screens.home.EmptyQueryException
-import pl.lambada.songsync.ui.screens.home.InternalErrorException
-import pl.lambada.songsync.ui.screens.home.NoTrackFoundException
+import pl.lambada.songsync.ui.screens.about.Providers
+import pl.lambada.songsync.util.EmptyQueryException
+import pl.lambada.songsync.util.InternalErrorException
+import pl.lambada.songsync.util.NoTrackFoundException
 import java.io.FileNotFoundException
 import java.net.UnknownHostException
