+++ app/src/main/java/pl/lambada/songsync/ui/screens/home/components/HomeAppBar.kt
@@ -7,7 +7,6 @@ import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Deselect
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material.icons.filled.SelectAll
-import androidx.compose.material3.Checkbox
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -26,9 +25,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import pl.lambada.songsync.R
-import pl.lambada.songsync.ui.screens.Providers
+import pl.lambada.songsync.ui.screens.about.Providers
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
