+++ app/src/main/java/pl/lambada/songsync/util/Exceptions.kt
@@ -1,0 +1,5 @@
+package pl.lambada.songsync.util
+
+class NoTrackFoundException : Exception()
+class InternalErrorException(msg: String) : Exception(msg)
+class EmptyQueryException : Exception()
